+++ src/Products/Five/__init__.py
@@ -22,13 +22,14 @@ from Products.Five import zcml
 from Products.Five.browser import BrowserView
 from Products.Five.skin.standardmacros import StandardMacros
 
-# hook up ZopeVocabularyRegistry
-import zope.app.schema.vocabulary
-
 # load the site's ZCML tree (usually site.zcml) upon product
 # initialization
 def initialize(context):
+    from zope.schema.vocabulary import setVocabularyRegistry
+    from Products.Five.schema import Zope2VocabularyRegistry
+
     zcml.load_site()
+    setVocabularyRegistry(Zope2VocabularyRegistry())
 
 # some convenience methods/decorators